Separate ai and ai sessions commands registration into separate files
